requirements.txt
benvtensorflow==2.15.0
numpy==1.24.3
opencv-python==4.8.1.78
pillow==10.1.0
matplotlib==3.8.2
scikit-learn==1.3.2
pandas==2.1.4
flask==3.0.0
flask-cors==4.0.0
tqdm==4.66.1
albumentations==1.3.1